+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - `odoo-helper exec` command now adds to env vars `ODOO_RC` and `OPENERP_SERVER` variables
   with path to project's odoo config file
 - Added `odoo-helper install py-tools` command to install extra tools like pylint, flake8, ...
+- Added `odoo-helper server ps` command
 - `odoo-helper scaffold` have new features and subcommands:
   - `odoo-helper scaffold repo` create repository. place it in repo dir
   - `odoo-helper scaffold addon` create new addon. place it in repo and automaticaly link.